pyproject.toml
[project]
name = "jinjatoyaml"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "pydantic>=2.13.3",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
